requirements.txt
requests
googlemaps
geopy
sentence_transformers
jina 
llama-index 
pymysql 
numpy
llama-index-vector-stores-tidbvector 
python-dotenv